backend/requirements.txt
django==4.2.25
dj_database_url==3.0.1
environs==14.3.0
django-debug-toolbar
django-phonenumber-field==8.3.0
djangorestframework
phonenumbers==8.13.30
geopy==2.4.1
requests==2.*
Pillow==12.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
rollbar==0.16.3
